+++ NEWS.md
@@ -3,6 +3,7 @@
 - allow string factors in meta data
 - make one-hot encoding the default for meta data variables with more than two
 categories, this avoids unreliable tests and improves interpretability (can be switched off with "make_onehot = false")
+- improve support for continuous meta data variables when using "fast=true"
 
 # v0.13.1 (latest)
 - fix stdlib dependency issue